+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-29 10:37:57
+
+### Changed Files
+- `src/features/policies/PolicyList.tsx`
+
+### Statistics
+```
+ 1 file changed, 61 insertions(+), 14 deletions(-)
+```
+
 ## [Checkpoint] - 2026-01-28 19:29:37
 
 ### Changed Files